Create a utils package
Move most of utils.go and the function WriteTOML there
