+++ buildpack/descriptor.go
@@ -2,6 +2,8 @@
 
 package buildpack
 
+import "github.com/BurntSushi/toml"
+
 type Descriptor struct {
 	API       string `toml:"api"`
 	Buildpack Info   `toml:"buildpack"`
@@ -33,6 +35,20 @@ type Order []Group
 
 type Group struct {
 	Group []GroupBuildpack `toml:"group"`
+}
+
+func ReadGroup(path string) (Group, error) {
+	var group Group
+	_, err := toml.DecodeFile(path, &group)
+	return group, err
+}
+
+func ReadOrder(path string) (Order, error) {
+	var order struct {
+		Order Order `toml:"order"`
+	}
+	_, err := toml.DecodeFile(path, &order)
+	return order.Order, err
 }
 
 func (bg Group) Append(group ...Group) Group {
