+++ buildpack/descriptor_test.go
@@ -1,0 +1,57 @@
+package buildpack_test
+
+import (
+	"io/ioutil"
+	"os"
+	"path/filepath"
+	"testing"
+
+	"github.com/google/go-cmp/cmp"
+	"github.com/sclevine/spec"
+	"github.com/sclevine/spec/report"
+
+	"github.com/buildpacks/lifecycle/buildpack"
+	h "github.com/buildpacks/lifecycle/testhelpers"
+)
+
+func TestUtils(t *testing.T) {
+	spec.Run(t, "Descriptor", testDescriptor, spec.Report(report.Terminal{}))
+}
+
+func testDescriptor(t *testing.T, when spec.G, it spec.S) {
+	when(".ReadOrder", func() {
+		var tmpDir string
+
+		it.Before(func() {
+			var err error
+			tmpDir, err = ioutil.TempDir("", "lifecycle.test")
+			if err != nil {
+				t.Fatal(err)
+			}
+		})
+
+		it.After(func() {
+			os.RemoveAll(tmpDir)
+		})
+
+		it("should return an ordering of buildpacks", func() {
+			h.Mkfile(t,
+				"[[order]]\n"+
+					`group = [{id = "A", version = "v1"}, {id = "B", optional = true}]`+"\n"+
+					"[[order]]\n"+
+					`group = [{id = "C"}, {}]`+"\n",
+				filepath.Join(tmpDir, "order.toml"),
+			)
+			actual, err := buildpack.ReadOrder(filepath.Join(tmpDir, "order.toml"))
+			if err != nil {
+				t.Fatalf("Unexpected error:\n%s\n", err)
+			}
+			if s := cmp.Diff(actual, buildpack.Order{
+				{Group: []buildpack.GroupBuildpack{{ID: "A", Version: "v1"}, {ID: "B", Optional: true}}},
+				{Group: []buildpack.GroupBuildpack{{ID: "C"}, {}}},
+			}); s != "" {
+				t.Fatalf("Unexpected list:\n%s\n", s)
+			}
+		})
+	})
+}
