+++ cache/image_cache.go
@@ -11,8 +11,8 @@ import (
 	"github.com/google/go-containerregistry/pkg/authn"
 	"github.com/pkg/errors"
 
-	"github.com/buildpacks/lifecycle"
 	"github.com/buildpacks/lifecycle/platform"
+	"github.com/buildpacks/lifecycle/utils"
 )
 
 const MetadataLabel = "io.buildpacks.lifecycle.cache.metadata"
@@ -74,7 +74,7 @@ func (c *ImageCache) SetMetadata(metadata platform.CacheMetadata) error {
 
 func (c *ImageCache) RetrieveMetadata() (platform.CacheMetadata, error) {
 	var meta platform.CacheMetadata
-	if err := lifecycle.DecodeLabel(c.origImage, MetadataLabel, &meta); err != nil {
+	if err := utils.DecodeLabel(c.origImage, MetadataLabel, &meta); err != nil {
 		return platform.CacheMetadata{}, nil
 	}
 	return meta, nil
