+++ cmd/lifecycle/analyzer.go
@@ -21,6 +21,7 @@ import (
 	"github.com/buildpacks/lifecycle/platform"
 	"github.com/buildpacks/lifecycle/platform/common"
 	"github.com/buildpacks/lifecycle/priv"
+	"github.com/buildpacks/lifecycle/utils"
 )
 
 type analyzeCmd struct {
@@ -186,7 +187,7 @@ func (a *analyzeCmd) Exec() error {
 		cacheStore lifecycle.Cache
 	)
 	if a.restoresLayerMetadata() {
-		group, err = lifecycle.ReadGroup(a.platform06.groupPath)
+		group, err = buildpack.ReadGroup(a.platform06.groupPath)
 		if err != nil {
 			return cmd.FailErr(err, "read buildpack group")
 		}
@@ -206,7 +207,7 @@ func (a *analyzeCmd) Exec() error {
 		return err
 	}
 
-	if err := buildpack.WriteTOML(a.analyzedPath, analyzedMD); err != nil {
+	if err := utils.WriteTOML(a.analyzedPath, analyzedMD); err != nil {
 		return errors.Wrap(err, "write analyzed.toml")
 	}
