+++ cmd/lifecycle/restorer.go
@@ -87,7 +87,7 @@ func (r *restoreCmd) Privileges() error {
 }
 
 func (r *restoreCmd) Exec() error {
-	group, err := lifecycle.ReadGroup(r.groupPath)
+	group, err := buildpack.ReadGroup(r.groupPath)
 	if err != nil {
 		return cmd.FailErr(err, "read buildpack group")
 	}
