+++ utils/utils_test.go
@@ -1,0 +1,78 @@
+package utils_test
+
+import (
+	"io/ioutil"
+	"os"
+	"path/filepath"
+	"testing"
+
+	"github.com/google/go-cmp/cmp"
+	"github.com/sclevine/spec"
+	"github.com/sclevine/spec/report"
+
+	"github.com/buildpacks/lifecycle/buildpack"
+	h "github.com/buildpacks/lifecycle/testhelpers"
+	"github.com/buildpacks/lifecycle/utils"
+)
+
+func TestUtils(t *testing.T) {
+	spec.Run(t, "Utils", testUtils, spec.Report(report.Terminal{}))
+}
+
+func testUtils(t *testing.T, when spec.G, it spec.S) {
+	when(".TruncateSha", func() {
+		it("should truncate the sha", func() {
+			actual := utils.TruncateSha("ed649d0a36b218c476b64d61f85027477ef5742045799f45c8c353562279065a")
+			if s := cmp.Diff(actual, "ed649d0a36b2"); s != "" {
+				t.Fatalf("Unexpected sha:\n%s\n", s)
+			}
+		})
+
+		it("should not truncate the sha with it's short", func() {
+			sha := "not-a-sha"
+			actual := utils.TruncateSha(sha)
+			if s := cmp.Diff(actual, sha); s != "" {
+				t.Fatalf("Unexpected sha:\n%s\n", s)
+			}
+		})
+
+		it("should remove the prefix", func() {
+			sha := "sha256:ed649d0a36b218c476b64d61f85027477ef5742045799f45c8c353562279065a"
+			actual := utils.TruncateSha(sha)
+			if s := cmp.Diff(actual, "ed649d0a36b2"); s != "" {
+				t.Fatalf("Unexpected sha:\n%s\n", s)
+			}
+		})
+	})
+
+	when(".WriteTOML", func() {
+		var tmpDir string
+
+		it.Before(func() {
+			var err error
+			tmpDir, err = ioutil.TempDir("", "lifecycle.test")
+			if err != nil {
+				t.Fatal(err)
+			}
+		})
+
+		it.After(func() {
+			os.RemoveAll(tmpDir)
+		})
+
+		it("should write TOML", func() {
+			group := buildpack.Group{Group: []buildpack.GroupBuildpack{{ID: "A", Version: "v1"}}}
+			if err := utils.WriteTOML(filepath.Join(tmpDir, "subdir", "group.toml"), group); err != nil {
+				t.Fatal(err)
+			}
+			b := h.Rdfile(t, filepath.Join(tmpDir, "subdir", "group.toml"))
+			if s := cmp.Diff(b,
+				"[[group]]\n"+
+					`  id = "A"`+"\n"+
+					`  version = "v1"`+"\n",
+			); s != "" {
+				t.Fatalf("Unexpected TOML:\n%s\n", s)
+			}
+		})
+	})
+}
